start cleaning up icons *sigh*
Gav, there is also kdelirc/icons which need to be cleaned up. Currently
you're installing all your icons globally which is only allowed for
app icons, not for actions or anything that's not available in kdelibs.
Those need to go into your app data folders; the code should pick that up
automatically like the global ones so you may want to check if things are
still working.

Another issue are the remotes or any data files that you install; those are
currently also installed in some obscure remotes/ folder which is incorrect.

Please contact me to sort out this mess. I'm going to move the kdelirc/icons
to the <prefix>/share/irkick/icons folder as well to get them out of the way of
the global folders; you may want to use kdelirc as a foldername in Makefile.am
depending on your requirements.

CCMAIL: gav@kde.org

svn path=/trunk/kdeutils/kdelirc/; revision=248509