BUGS:103668 A slightly more comprehensive fix. Patched the readLine method so it only returns null when there actually is no data coming, rather than when the buffer is temporarily empty. It still uses thew ugly timeout though :-(
As rizsanyi points out, serious code reorganisation is necessary to fix
this properly---probably to change the comms stuff so it's event driven.
Not a nice task.

Though it compiles, this fix is as yet untested, since I don't have the
appropriate system on which to test it (though it does compile). Feedback
most welcome.

svn path=/trunk/kdeutils/kdelirc/; revision=406070